engine/pyproject.toml
[project]
name = "igaming-operation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.35",
    "google-genai>=1.60.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.2.0",
    "gspread>=6.2.1",
    "httpx[socks]>=0.28.1",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.1",
    "sshtunnel>=0.4.0",
    "tabulate>=0.9.0",
    "playwright>=1.57.0",
    "pyotp>=2.9.0",
    "telethon>=1.42.0",
    "prompt-toolkit>=3.0.52",
    "openai>=2.16.0",
]
